Refactor: News 기능으로 분리
