+++ src/features/news/index.ts
@@ -1,0 +1,1 @@
+export { News } from "./ui/News";
